File consolidation, everything is now locked behind login screen (to be fixed later), database connections now work
