+++ src/main/java/com/costi/csw9/Model/User.java
@@ -19,9 +19,9 @@ import java.util.Collections;
 @Entity
 public class User implements UserDetails {
 
-    @Id
-    @SequenceGenerator(name = "user_sequence", sequenceName = "user_sequence", allocationSize = 1)
-    @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "user_sequence")
+   @Id
+   @SequenceGenerator(name = "user_sequence", sequenceName = "user_sequence", allocationSize = 1)
+   @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "user_sequence")
    private Long id;
    String firstName;
    String lastName;
