+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url = jdbc:h2:file:C:/Users/costi_dmi8dsd/OneDrive/Documents/Costi Journaling/InteliJ Workspace/Costi Survival World 9 Website/Data/Database
+spring.datasource.url = jdbc:h2:tcp://localhost/~/OneDrive/Documents/Costi Journaling/InteliJ Workspace/Costi Survival World 9 Website/Data/Database
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
