+++ src/main/resources/templates/main/Projects.html
@@ -6,7 +6,7 @@
 
 <div class="container-fluid row px-0 mx-0 d-none d-lg-block">
     <div class="col-lg-3 bg-dark px-5 position-fixed" style="height:100vh;padding-top:70px">
-        <h1 class="fs-1">Development Eras</h1>
+        <h2 class="fs-2">Development Eras</h2>
         <ul class="nav flex-column">
             <li class="nav-item">
                 <h4 class="fs-4"><a class="nav-link" href="#IDE">Interdimensional Era (current)</a></h4>
