+++ src/test/java/com/costi/csw9/CostiOnlineStarterTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class CostiSurvivalWorld9WebsiteApplicationTests {
+class CostiOnlineStarterTests {
 
 	@Test
 	void contextLoads() {
